Review fixes: a diagnostic that could end in silence, and two wrong numbers
Five code findings, all verified by reading the source, plus two published
figures that did not survive re-derivation from the raw data.

Code:

- The I2C-bus-failure path skipped the whole gate section and fell through to
  the closing rule with no verdict, which reads exactly like a pass. It is
  reachable: board_variant_detect() only WARNS when its own
  i2c_del_master_bus() fails (board_variant.c:112-115), leaving I2C_NUM_0
  occupied so our open returns ESP_ERR_INVALID_STATE. Now prints an explicit
  STOP saying the run proved nothing.
- esp_flash_get_size() and esp_read_mac() returns were discarded, so a failed
  read printed "flash: 0 bytes (0 MB)" — a value, not a failure. AC3 records
  that number as OBSERVED and resolving 8 MB vs 16 MB is half this app's job.
  Both now report READ FAILED.
- The D-005 ruling printed before the bus scan that corroborates it, so a
  transcript could read "D-005 HOLDS" directly above "AC2 FAIL". report_variant
  now runs after corroboration and withholds the ruling without it.
- A local named ac12 held the AC1/AC2 result, one line above the actual AC12
  PMU gate. Renamed variant_ok.
- Dropped an unused string.h; cast the promoted uint16_t voltage args at the
  six %u sites.

Published figures:

- "the same regions differ 57-95%" excluded the partition table's 2.7%. Two
  ESP-IDF partition tables are mostly identical padding, so that figure is
  expected and unremarkable — but a range that silently drops its own minimum
  is not a measurement. All five regions now stated individually.
- "CST816S (14)" and "CO5300 (31)" were grep -ci totals across every casing,
  attributed to one capitalisation each. Case-sensitive: CO5300 1, co5300 30,
  CST816S 2, cst816 12. The V1 parts are zero in any casing, so the conclusion
  stands, but the numbers as written would not have reproduced.

Build clean, AC8 still 0 with the positive control still 2.

+++ firmware/board_check/main/pmu_read.c
@@ -2,7 +2,6 @@
 
 #include <stdbool.h>
 #include <stdio.h>
-#include <string.h>
 
 #include "esp_log.h"
 
@@ -151,15 +150,15 @@ pmu_status_t pmu_dump(i2c_master_bus_handle_t bus)
         }
     }
     printf("0x%02X DCDC1_VOL    raw=0x%02X   %s  ~%u mV\n", REG_DC_VOL0, dcv[0],
-           (dc_on & 0x01) ? "ON " : "off", dc1_mv(dcv[0]));
+           (dc_on & 0x01) ? "ON " : "off", (unsigned)dc1_mv(dcv[0]));
     printf("0x%02X DCDC2_VOL    raw=0x%02X   %s  ~%u mV\n", REG_DC_VOL1, dcv[1],
-           (dc_on & 0x02) ? "ON " : "off", dc234_mv(dcv[1]));
+           (dc_on & 0x02) ? "ON " : "off", (unsigned)dc234_mv(dcv[1]));
     printf("0x%02X DCDC3_VOL    raw=0x%02X   %s  ~%u mV\n", REG_DC_VOL2, dcv[2],
-           (dc_on & 0x04) ? "ON " : "off", dc3_mv(dcv[2]));
+           (dc_on & 0x04) ? "ON " : "off", (unsigned)dc3_mv(dcv[2]));
     printf("0x%02X DCDC4_VOL    raw=0x%02X   %s  ~%u mV\n", REG_DC_VOL3, dcv[3],
-           (dc_on & 0x08) ? "ON " : "off", dc234_mv(dcv[3]));
+           (dc_on & 0x08) ? "ON " : "off", (unsigned)dc234_mv(dcv[3]));
     printf("0x%02X DCDC5_VOL    raw=0x%02X   %s  ~%u mV\n", REG_DC_VOL4, dcv[4],
-           (dc_on & 0x10) ? "ON " : "off", dc5_mv(dcv[4]));
+           (dc_on & 0x10) ? "ON " : "off", (unsigned)dc5_mv(dcv[4]));
 
     /* --- LDO block ------------------------------------------------------- */
     uint8_t l0 = 0, l1 = 0;
@@ -183,7 +182,7 @@ pmu_status_t pmu_dump(i2c_master_bus_handle_t bus)
             status = PMU_AMBIGUOUS;
         }
         bool on = (i <= 7) ? !!(l0 & (1u << i)) : !!(l1 & 0x01);
-        uint16_t mv = (i == 6) ? cpusldo_mv(v) : ldo100_mv(v);
+        unsigned mv = (i == 6) ? cpusldo_mv(v) : ldo100_mv(v);
         printf("0x%02X %-7s VOL  raw=0x%02X   %s  ~%u mV\n",
                reg, ldo_name[i], v, on ? "ON " : "off", mv);
     }
